+++ CHANGELOG.md
@@ -1,3 +1,10 @@
+## 0.0.2
+* **BREAK**: `StateAtom().stream` no longer emits the initial state on subscription.
+* Tests added for `Atom`s
+* Tests added for `SingleBond`s
+* Tests added for `Molecule`s
+
+
 ## 0.0.1
 
 * Initial release